Fix warning: Comparison of unsigned expression < 0 is always false.

+++ src/ConEmu/Macro.cpp
@@ -1121,7 +1121,7 @@ LPWSTR ConEmuMacro::ExecuteMacro(LPWSTR asMacro, CRealConsole* apRCon, bool abFr
 
 LRESULT ConEmuMacro::ExecuteMacroSync(WPARAM wParam, LPARAM lParam)
 {
-	if (wParam < 0 || wParam >= countof(Functions))
+	if ((int)wParam < 0 || wParam >= countof(Functions))
 	{
 		_ASSERTE(wParam >= 0 && wParam < countof(Functions));
 		return 0;
